requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.2
django-crontab==0.7.1
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
django-timezone-field==5.1
djangorestframework==3.14.0
flower==1.1.0
humanize==4.7.0
idna==3.4
kombu==5.3.1
phonenumbers==8.13.15
prometheus-client==0.17.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tornado==6.3.2
typing_extensions==4.7.0
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
